Autocorrect Style/IfUnlessModifier

+++ lib/alexandria/models/library.rb
@@ -363,9 +363,7 @@ module Alexandria
 
     def self.canonicalise_isbn(isbn)
       numbers = extract_numbers(isbn)
-      if valid_ean?(isbn) && (numbers[0..2] != [9, 7, 8])
-        return isbn
-      end
+      return isbn if valid_ean?(isbn) && (numbers[0..2] != [9, 7, 8])
       canonical = if valid_ean?(isbn)
                     # Looks like an EAN number -- extract the intersting part and
                     # calculate a checksum. It would be nice if we could validate
@@ -390,9 +388,7 @@ module Alexandria
       # Let's initialize the saved identifier if not already
       # (backward compatibility from 0.4.0)
       # book.saved_ident ||= book.ident
-      if book.saved_ident.nil? || book.saved_ident.empty?
-        book.saved_ident = book.ident
-      end
+      book.saved_ident = book.ident if book.saved_ident.nil? || book.saved_ident.empty?
       if book.ident != book.saved_ident
         # log.debug { "Backwards compatibility step: #{book.saved_ident.inspect}, #{book.ident.inspect}" }
         FileUtils.rm(yaml(book.saved_ident))
@@ -465,9 +461,7 @@ module Alexandria
         end
 
         # Remove the file if its blank.
-        if Alexandria::UI::Icons.blank?(cover_file)
-          File.delete(cover_file)
-        end
+        File.delete(cover_file) if Alexandria::UI::Icons.blank?(cover_file)
       end
     end
